refactored delete user, moved logic to service

+++ domain/services/userService.js
@@ -36,33 +36,39 @@ var userService = {
       }
     })
   },
-  deleteUser: (userData, callback) => {
-    userRepository.delete(userData, (error) => {
-      if (!error) {
-        var userCheckIds = typeof(userData.checks) == 'object' && userData.checks instanceof Array ? userData.checks : []
-        var nChecksToDelete = userCheckIds.length
-        if (0 < nChecksToDelete) {
-          var nChecksDeleted = 0
-          var deletionErrors = []
-          userCheckIds.forEach((checkId) => {
-            checkRepository.delete(checkId, (error) => {
-              if (error) {
-                deletionErrors.push(error)
-              }
-              nChecksDeleted++;
-              if (nChecksDeleted == nChecksToDelete) {
-                if (!deletionErrors.length) {
-                  callback(false)
-                } else {
-                  callback({ deletionErrors })
-                }
-              }
-
-            })
-          })
-        } else {
-          callback(false)
-        }
+  deleteUser: (phone, callback) => {
+    userRepository.read(phone, (error, userData) => {
+      if (!error && userData) {
+        userRepository.delete(userData.phone, (error) => {
+          if (!error) {
+            var userCheckIds = typeof(userData.checks) == 'object' && userData.checks instanceof Array ? userData.checks : []
+            var nChecksToDelete = userCheckIds.length
+            if (0 < nChecksToDelete) {
+              var nChecksDeleted = 0
+              var deletionErrors = []
+              userCheckIds.forEach((checkId) => {
+                checkRepository.delete(checkId, (error) => {
+                  if (error) {
+                    deletionErrors.push(error)
+                  }
+                  nChecksDeleted++;
+                  if (nChecksDeleted == nChecksToDelete) {
+                    if (!deletionErrors.length) {
+                      callback(false)
+                    } else {
+                      callback({ deletionErrors })
+                    }
+                  }
+    
+                })
+              })
+            } else {
+              callback(false)
+            }
+          } else {
+            callback(error)
+          }
+        })
       } else {
         callback(error)
       }
